+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+* **2.1.1**
+    - [#63](https://github.com/joecorcoran/judge/pull/63): add array length validation
+    - [#68](https://github.com/joecorcoran/judge/pull/68): count newlines as two characters when validating length
+* **2.1.0**
+    - [#55](https://github.com/joecorcoran/judge/pull/55): move confirmation validator to comfirmation field
+    - [#56](https://github.com/joecorcoran/judge/pull/56): add `email_field` support
+    - [#58](https://github.com/joecorcoran/judge/pull/58): add presence validation support to radio buttons
+    - [#60](https://github.com/joecorcoran/judge/pull/60): support Ruby-style regex in JavaScript
+* **2.0.6**
+    - [#46](https://github.com/joecorcoran/judge/pull/46): remove Responders
+    - [#49](https://github.com/joecorcoran/judge/pull/49): remove status check from tryClose
+    - [#48](https://github.com/joecorcoran/judge/pull/48): added original value to uniqueness validations
+    - [#51](https://github.com/joecorcoran/judge/pull/51): Seperate required parameters from conditional parameters
+    - [#52](https://github.com/joecorcoran/judge/pull/52): Fix nested model uniqueness validation
+* **2.0.5**
 * **2.0.4** Better empty param handling and safer loading of engine files.
 * **2.0.3** Fixed Internet Explorer XHR bug.
 * **2.0.2** Fixed controller inheritance bug.